+++ Controllers/TripsController.cs
@@ -178,10 +178,8 @@ namespace mr_shtrahman.Controllers
             {
                 return NotFound();
             }
-            else
-            {
                 await DeleteConfirmed(id);
-            }
+            
 
             return RedirectToAction(nameof(Index));
         }
@@ -189,8 +187,13 @@ namespace mr_shtrahman.Controllers
         // POST: Trips/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int? id)
         {
+            if (id == null)
+            {
+                return NotFound();
+            }
+
             var trip = await _context.Trip.FindAsync(id);
             _context.Trip.Remove(trip);
             await _context.SaveChangesAsync();
